+++ .idea/sqldialects.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/spring-boot/src/main/java/springboot/databaseSetup.java" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/spring-boot/src/main/java/springboot/database/Setup.java" dialect="GenericSQL" />
   </component>
 </project>
