+++ spring-boot/src/main/java/springboot/Application.java
@@ -2,13 +2,13 @@ package springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import springboot.database.Setup;
 
 @SpringBootApplication
 public class Application {
 
     public static void main(String[] args) {
-        databaseSetup.initialize();
-
+        Setup.initialize();
         SpringApplication.run(Application.class, args);
     }
 }
